Fix!: Volume parameter responsiveness
The headphone volume parameter is now as responsive as regular parameters
This changes the mapping of the headphonevol system parameter

+++ sw/Core/Src/plinky/hardware/codec.h
@@ -5,4 +5,4 @@ void HAL_SAI_RxCpltCallback(SAI_HandleTypeDef* hi2s);
 void HAL_SAI_RxHalfCpltCallback(SAI_HandleTypeDef* hi2s);
 
 void codec_init(void);
-void codec_set_volume(u16 vol);
+void codec_set_volume(u8 vol);
